+++ source/RTSCamera/src/Patch/Naval/Patch_MissionGauntletNavalOrderUIHandler.cs
@@ -1,5 +1,6 @@
 ﻿using HarmonyLib;
 using MissionSharedLibrary.Utilities;
+using RTSCamera.Patch.Fix;
 using System;
 using System.Reflection;
 using TaleWorlds.MountAndBlade;
@@ -40,6 +41,9 @@ namespace RTSCamera.Patch.Naval
             MissionFormationTargetSelectionHandler ____formationTargetHandler,
             Object ____shipTargetHandler)
         {
+            // MissionOrderVM.OnTroopFormationSelected will close the order UI and open it again to refresh the available orders.
+            // We should allow it to be closed so that orders will be refreshed. Or the advance visual orders may not be available when switching from agent view to rts view.
+            Patch_MissionOrderVM.AllowClosingOrderUI = true;
             bool isDisabled = !Patch_NavalDLCHelpers.IsShipOrderAvailable();
             ____formationTargetHandler?.SetIsFormationTargetingDisabled(isDisabled);
             if (____shipTargetHandler == null)