+++ source/RTSCamera/src/RTSCameraSubModule.cs
@@ -117,6 +117,7 @@ namespace RTSCamera
                     _successPatch &= Patch_NavalDLCHelpers.Patch(_harmony);
                     _successPatch &= Patch_MissionGauntletNavalOrderUIHandler.Patch(_harmony);
                     _successPatch &= Patch_NarvalShipTargetSelectionHandler.Patch(_harmony);
+                    _successPatch &= Patch_ShipAgentSpawnLogicTeamSide.Patch(_harmony);
                 }
 
                 // Use Patch to add game menu